practice/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
comtypes==1.1.10
constants==0.6.0
dataclasses==0.8
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==3.2.5
django-allauth==0.45.0
django-baton==2.2.1
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djongo==1.3.6
node==0.9.26
nodeenv==1.6.0
PyJWT==1.7.1
pymongo==3.12.0
requests==2.26.0

